# docs/endpoints/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ithina Admin Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/stores": {
      "get": {
        "tags": [
          "stores"
        ],
        "summary": "List Stores",
        "description": "List stores visible to the caller, paginated.\n\nPLATFORM sees all rows (D-29 OR-branch); TENANT sees own-tenant\nonly. ``pagination.total`` is RLS-filtered (reflects what the\ncaller can see, not the platform total).\n\nSearch is trimmed; empty after trim is treated as no filter.",
        "operationId": "list_stores_api_v1_stores_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a single tenant. Useful for PLATFORM callers scoping a list view; TENANT callers see only their own tenant regardless (RLS handles it).",
              "title": "Tenant Id"
            },
            "description": "Filter to a single tenant. Useful for PLATFORM callers scoping a list view; TENANT callers see only their own tenant regardless (RLS handles it)."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StoreStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: OPENING, ACTIVE, INACTIVE, or CLOSED.",
              "title": "Status"
            },
            "description": "Filter by status: OPENING, ACTIVE, INACTIVE, or CLOSED."
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact-match filter on country (case-sensitive).",
              "title": "Country"
            },
            "description": "Exact-match filter on country (case-sensitive)."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring match across name and store_code.",
              "title": "Search"
            },
            "description": "Case-insensitive substring match across name and store_code."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort key. Field-based: name_asc, name_desc, created_at_asc, created_at_desc, status_asc, country_asc. Cross-table: tenant_name_asc (default), tenant_name_desc; both apply a stable secondary sort by stores.name ASC for deterministic pagination within a tenant. Unknown values -> 400 INVALID_SORT_KEY.",
              "default": "tenant_name_asc",
              "title": "Sort"
            },
            "description": "Sort key. Field-based: name_asc, name_desc, created_at_asc, created_at_desc, status_asc, country_asc. Cross-table: tenant_name_asc (default), tenant_name_desc; both apply a stable secondary sort by stores.name ASC for deterministic pagination within a tenant. Unknown values -> 400 INVALID_SORT_KEY."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Standard pagination.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Standard pagination."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Standard pagination.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Standard pagination."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stores"
        ],
        "summary": "Create Store",
        "description": "Create a new store. Multi-audience.\n\nPLATFORM tier (Super Admin, Platform Admin) creates for any tenant\nvia the GLOBAL->TENANT cascade. TENANT tier (Owner) creates for own\ntenant via direct ``.TENANT`` grant; ``tenant_id`` in body verified\nagainst the caller's RLS-bound session — a cross-tenant id is\ninvisible to TENANT callers and surfaces as 404\n``ORG_NODE_NOT_FOR_STORE`` (when ``org_node_id`` provided) or as a\nDB FK violation otherwise.\n\nServer omits ``status`` so the DDL default fires (LD8). The DDL\ndefault in v0 is ``ACTIVE`` (see ``stores_v5.sql``); the prompt's\nintent (\"via DDL default\") is honoured by omitting the column\nfrom the INSERT — if a future migration changes the default to\n``OPENING`` (the product-intended initial state implied by the\nlifecycle enum ordering), no app code change is required.\nAudit-actor pairs populate from ``auth.user_id`` and the\naudience-bridged ``ActorUserType``.\n\n``store_code`` uniqueness checked per-tenant (case-insensitive,\naligns with DDL partial unique index) before INSERT — pre-check\nsurfaces as 409 ``DUPLICATE_STORE_CODE``. Future NOT NULL migration\non ``store_code`` / ``tax_treatment`` will tighten DDL alongside\nthese app-layer schema constraints.\n\n``parent_org_node_id`` is REQUIRED (Step 6.21.2). It must reference\na non-STORE org_node in the same tenant. The server creates the\npaired STORE-type org_node under it inside the same transaction as\nthe ``stores`` row. 404 ``PARENT_NODE_NOT_FOUND`` for missing or\nRLS-filtered parents; 422 ``INVALID_PARENT_NODE_TYPE`` for\nSTORE-type parents; 409 ``DUPLICATE_ORG_NODE_CODE`` when the\nstore_code collides with any other org_node code in the tenant.",
        "operationId": "create_store_api_v1_stores_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/{store_id}": {
      "get": {
        "tags": [
          "stores"
        ],
        "summary": "Get Store",
        "description": "Return the full detail shape for a single store.\n\n404 if the row doesn't exist OR is RLS-filtered. Per D-17 the\nhandler can't and shouldn't distinguish — both produce the same\nbody shape (``code: STORE_NOT_FOUND``). The anchor dep\n``get_store_anchor`` also raises ``StoreNotFoundError`` on the\nsame conditions, so cross-tenant probes hit 404 ahead of the gate\nbody fire path.",
        "operationId": "get_store_api_v1_stores__store_id__get",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "stores"
        ],
        "summary": "Patch Store",
        "description": "Partial update of a store. Multi-audience.\n\nMutable fields: ``name``, ``store_code``, ``country``, ``timezone``,\n``currency``, ``tax_treatment``, ``address``, ``latitude``,\n``longitude``, and ``parent_org_node_id`` (Step 6.21.2). Other\nfields rejected at the schema layer via ``extra=\"forbid\"`` —\nincluding ``status`` (lifecycle is in Step 6.17.4's\n``/change_status``), ``tenant_id``, and ``org_node_id`` (the\nstore's slot id is immutable; reparent uses\n``parent_org_node_id``).\n\nStep 6.21.2 cascade: ``name``, ``store_code``, and\n``parent_org_node_id`` changes propagate to the paired STORE-type\norg_node atomically inside one transaction. Org_node-side\nUNIQUE collisions on ``store_code`` change surface as 409\n``DUPLICATE_ORG_NODE_CODE`` (broader scope than the\nstores-only ``DUPLICATE_STORE_CODE`` pre-check). Parent\nvalidation produces 404 ``PARENT_NODE_NOT_FOUND`` or 422\n``INVALID_PARENT_NODE_TYPE`` for invalid\n``parent_org_node_id`` values.\n\nBehaviour:\n  - 422 ``EMPTY_PATCH`` when no fields are set.\n  - 409 ``DUPLICATE_STORE_CODE`` on rename to a value held by\n    another store in the same tenant (case-insensitive). Rename to\n    the same value (after case-fold) is a 200 no-op.\n  - 404 ``STORE_NOT_FOUND`` when the row is missing or RLS-filtered\n    (RLS-as-404 per D-17). Anchor dep ``get_store_anchor`` fires\n    the same code via the gate path for cross-tenant probes.\n  - ``updated_at`` bumps on every non-empty PATCH (LD4); a\n    non-empty same-as-current PATCH still bumps.",
        "operationId": "patch_store_api_v1_stores__store_id__patch",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Store Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/{store_id}/set-status": {
      "post": {
        "tags": [
          "stores"
        ],
        "summary": "Set Store Status",
        "description": "State-transition endpoint for a store. Multi-audience.\n\nGate: same as PATCH ``/stores/{store_id}`` —\n``ADMIN.STORES.CONFIGURE.TENANT`` with ``anchor_dep=get_store_anchor``.\nSUPER_ADMIN + PLATFORM_ADMIN pass via the GLOBAL->TENANT cascade;\nTENANT OWNER passes via the direct ``.TENANT`` grant.\n\n9-cell liberal transition matrix per LD1: all transitions allowed\nEXCEPT ``*->OPENING``. CLOSED is reversible. Same-state returns\n409 ``INVALID_STATE_TRANSITION`` (LD5; mirrors tenants\n``allowed_sources`` convention).\n\nBehaviour:\n  - 200 + full ``StoreDetail`` on successful transition. Class 1\n    (into-CLOSED) populates ``closed_*``; Class 2 (out-of-CLOSED)\n    nulls it; Class 3 (between non-CLOSED) leaves it untouched.\n  - 404 ``STORE_NOT_FOUND`` when the store id is missing or\n    RLS-filtered. Anchor dep ``get_store_anchor`` fires first on\n    cross-tenant probes by TENANT callers.\n  - 409 ``INVALID_STATE_TRANSITION`` on rejected transitions\n    (including same-state). Context kwargs ``store_id`` +\n    ``target_status`` reach ``exc.context`` for logs; response\n    envelope ``details`` stays ``null`` per the Q7 lock.\n\n``reason`` is forward-compatible (LD3). It is consumed by Pydantic\nvalidation but NOT passed to the repo in this step. When Step 6.2\naudit_log ships, the handler gains an\n``audit_log_repo.write(...reason=body.reason)`` call after the\nrepo call succeeds; no API change required.",
        "operationId": "set_store_status_api_v1_stores__store_id__set_status_post",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Store Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreSetStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StoreDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "tenant_name": {
            "type": "string",
            "title": "Tenant Name"
          },
          "org_node_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Node Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "store_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Code"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "tax_treatment": {
            "$ref": "#/components/schemas/TaxTreatment"
          },
          "status": {
            "$ref": "#/components/schemas/StoreStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "tenant_name",
          "org_node_id",
          "name",
          "store_code",
          "country",
          "timezone",
          "address",
          "latitude",
          "longitude",
          "currency",
          "tax_treatment",
          "status",
          "created_at",
          "updated_at",
          "closed_at"
        ],
        "title": "StoreDetail",
        "description": "Detail response shape for ``GET /api/v1/stores/{store_id}``.\n\n17 fields. Audit-actor IDs (Pattern (b) per D-13) deliberately\nhidden; ``org_node_id`` exposed as a bare UUID column (locked\ndecision 8) — frontend resolves the name via ``/org-tree`` if\nneeded."
      },
...
      "StoreListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StoreListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "StoreListResponse",
        "description": "List endpoint response envelope: ``{items, pagination}`` per D-30."
      },
...
      "StoreStatus": {
        "type": "string",
        "enum": [
          "OPENING",
          "ACTIVE",
          "INACTIVE",
          "CLOSED"
        ],
        "title": "StoreStatus",
        "description": "Lifecycle states for a store. Mirrors ``store_status_enum``."
      },
...
    }
  }
}